chore: add mote tests
