+++ .github/workflows/flutter_ci.yaml
@@ -398,12 +398,6 @@ jobs:
           tar -xf appflowy_flutter.tar.gz
           ls -al
 
-      - name: Enable localhost env
-        working-directory: frontend/appflowy_flutter
-        run: |
-          echo 'APPFLOWY_CLOUD_URL=http://localhost' >> .env
-          dart run build_runner clean && dart run build_runner build --delete-conflicting-outputs
-
       - name: Run flutter pub get
         working-directory: frontend
         run: cargo make pub_get
