+++ frontend/appflowy_flutter/integration_test/settings/settings_runner.dart
@@ -1,13 +1,11 @@
 import 'package:integration_test/integration_test.dart';
 
 import 'notifications_settings_test.dart' as notifications_settings_test;
-import 'user_icon_test.dart' as user_icon_test;
 import 'user_language_test.dart' as user_language_test;
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
 
   notifications_settings_test.main();
-  user_icon_test.main();
   user_language_test.main();
 }
