+++ frontend/appflowy_flutter/integration_test/sync/user_setting_sync_test.dart
@@ -1,0 +1,95 @@
+// ignore_for_file: unused_import
+
+import 'dart:io';
+
+import 'package:appflowy/env/cloud_env.dart';
+import 'package:appflowy/generated/locale_keys.g.dart';
+import 'package:appflowy/startup/startup.dart';
+import 'package:appflowy/user/application/auth/af_cloud_mock_auth_service.dart';
+import 'package:appflowy/user/application/auth/auth_service.dart';
+import 'package:appflowy/workspace/application/settings/prelude.dart';
+import 'package:appflowy/workspace/presentation/settings/widgets/setting_appflowy_cloud.dart';
+import 'package:appflowy/workspace/presentation/settings/widgets/settings_user_view.dart';
+import 'package:appflowy/workspace/presentation/widgets/user_avatar.dart';
+import 'package:appflowy_backend/protobuf/flowy-folder2/view.pb.dart';
+import 'package:easy_localization/easy_localization.dart';
+import 'package:flowy_infra/uuid.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_test/flutter_test.dart';
+import 'package:path/path.dart' as p;
+import 'package:integration_test/integration_test.dart';
+import '../util/dir.dart';
+import '../util/emoji.dart';
+import '../util/mock/mock_file_picker.dart';
+import '../util/util.dart';
+
+void main() {
+  IntegrationTestWidgetsFlutterBinding.ensureInitialized();
+  final email = '${uuid()}@appflowy.io';
+  const name = 'nathan';
+
+  group('appflowy cloud setting', () {
+    testWidgets('sync user name and icon to server', (tester) async {
+      await tester.initializeAppFlowy(
+        cloudType: AuthenticatorType.appflowyCloud,
+        userEmail: email,
+      );
+      await tester.tapGoogleLoginInButton();
+      tester.expectToSeeHomePage();
+
+      await tester.openSettings();
+      await tester.openSettingsPage(SettingsPage.user);
+      final userAvatarFinder = find.descendant(
+        of: find.byType(SettingsUserView),
+        matching: find.byType(UserAvatar),
+      );
+
+      // Open icon picker dialog and select emoji
+      await tester.tap(userAvatarFinder);
+      await tester.pumpAndSettle();
+      await tester.tapEmoji('😁');
+      await tester.pumpAndSettle();
+      final UserAvatar userAvatar =
+          tester.widget(userAvatarFinder) as UserAvatar;
+      expect(userAvatar.iconUrl, '😁');
+
+      // enter user name
+      final userNameFinder = find.descendant(
+        of: find.byType(SettingsUserView),
+        matching: find.byType(UserNameInput),
+      );
+      await tester.enterText(userNameFinder, name);
+      await tester.pumpAndSettle();
+    });
+  });
+
+  testWidgets('get user icon and name from server', (tester) async {
+    await tester.initializeAppFlowy(
+      cloudType: AuthenticatorType.appflowyCloud,
+      userEmail: email,
+    );
+    await tester.tapGoogleLoginInButton();
+    tester.expectToSeeHomePage();
+    await tester.pumpAndSettle();
+
+    await tester.openSettings();
+    await tester.openSettingsPage(SettingsPage.user);
+
+    // verify icon
+    final userAvatarFinder = find.descendant(
+      of: find.byType(SettingsUserView),
+      matching: find.byType(UserAvatar),
+    );
+    final UserAvatar userAvatar = tester.widget(userAvatarFinder) as UserAvatar;
+    expect(userAvatar.iconUrl, '😁');
+
+    // verify name
+    final userNameFinder = find.descendant(
+      of: find.byType(UserNameInput),
+      matching: find.byType(TextField),
+    );
+    final UserNameInput userNameInput =
+        tester.widget(userNameFinder) as UserNameInput;
+    expect(userNameInput.name, name);
+  });
+}
